requirements.txt
streamlit>=1.35.0
pandas>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
google-genai>=1.0.0
plotly>=5.0.0
networkx>=3.0
beautifulsoup4>=4.12.0
python-whois>=0.9.0
